requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-allauth==64.1.0
django-braces==1.15.0
django-debug-toolbar==4.4.6
django-dotenv==1.4.2
django-model-utils==4.5.1
django-ratelimit==4.1.0
django-requests==0.0.2
django-rest-framework==0.1.0
django-star-ratings==0.9.2
djangorestframework==3.15.2
idna==3.7
jwt==1.3.1
load-dotenv==0.1.0
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
qrcode==8.0
regex==2024.7.24
requests==2.32.3
sqlparse==0.5.1
swapper==1.4.0
Unidecode==1.3.8
urllib3==2.2.2
elasticsearch==8.12.0
